+++ lib/bootstrap.scm
@@ -702,15 +702,15 @@
   (and (symbol? value) (--> value (is_gensym))))
 
 ;; ---------------------------------------------------------------------------------------
-(define (degree->radian x)
-  "(degree->radian x)
+(define (degree->radians x)
+  "(degree->radians x)
 
    Convert degree to radians."
   (* x (/ Math.PI 180)))
 
 ;; ---------------------------------------------------------------------------------------
-(define (radian->degree x)
-  "(degree->radian x)
+(define (radians->degree x)
+  "(radians->degree x)
 
    Convert radians to degree."
   (* x (/ 180 Math.PI)))